variance

+++ variance.py
@@ -1,10 +1,10 @@
 def mean(a):
-    return sum(a) / len(a)
+    return float(sum(a) / len(a))
 
 def variance(a):
     x = mean(a)
     variance = sum([(i-x)**2 for i in a])
-    return variance
+    return float(variance)
 
 y = list(map(int,input().split()))
 z= variance(y)
